refactor submission parsing utilities out of aws_keyspaces for reusal
